Correction de bugs.
svn path=/java/trunk/; revision=1285

+++ sae/sae-ecde/src/main/java/fr/urssaf/image/sae/ecde/service/strategy/impl/JaxbResultatSerializer.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import fr.urssaf.image.sae.bo.model.MetadataError;
 import fr.urssaf.image.sae.bo.model.SAEError;
 import fr.urssaf.image.sae.bo.model.untyped.UntypedDocumentOnError;
+import fr.urssaf.image.sae.ecde.exception.EcdeRuntimeException;
 import fr.urssaf.image.sae.ecde.exception.EcdeXsdException;
 import fr.urssaf.image.sae.ecde.modele.commun_sommaire_et_resultat.BatchModeType;
 import fr.urssaf.image.sae.ecde.modele.commun_sommaire_et_resultat.ErreurType;
@@ -128,8 +129,16 @@ public class JaxbResultatSerializer implements ResultatSerializerStrategy {
    private void createFlag(String ecdeDirectory) throws IOException {
       try {
          File flag = new File(FilenameUtils.concat(ecdeDirectory,"fin_traitement.flag"));
-         flag.delete();
-         flag.createNewFile();
+         if (flag.exists()) {
+            boolean delete = flag.delete();
+            if (!delete) {
+               throw new EcdeRuntimeException("Erreur lors de la suppression du fichier flag!");
+            }
+         }   
+         boolean create = flag.createNewFile();
+         if (!create) {
+            throw new EcdeRuntimeException("Erreur lors de l'ecriture du fichier flag!");
+         }
       } catch (IOException except) {
          throw new IOException("Erreur de création du fichier flag.", except);
       }
